requirements.txt
-e git+https://github.com/django/django.git#egg=django1.5
-e git+https://github.com/j2a/pytils.git#egg=django-pytils
South==0.7.6
sorl-thumbnail==11.12
django-tagging==0.3.1
django-registration==0.8
django-pagination==1.0.7
django-sorting==0.1